fix: change telegram_user_id from Integer to BigInteger to support large Telegram user IDs

+++ api/models/user.py
@@ -1,6 +1,6 @@
 """User ORM model and Pydantic schemas."""
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, DateTime, Index
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Index
 from sqlalchemy.orm import relationship
 from api.core.db import Base
 
@@ -10,7 +10,7 @@ class User(Base):
     __tablename__ = "users"
     
     id = Column(Integer, primary_key=True, index=True)
-    telegram_user_id = Column(Integer, unique=True, nullable=False, index=True)
+    telegram_user_id = Column(BigInteger, unique=True, nullable=False, index=True)
     username = Column(String, nullable=True)
     first_name = Column(String, nullable=True)
     last_name = Column(String, nullable=True)
